refactor(cf): separate SG details sections
