+++ app/scripts/modules/cloudfoundry/src/serverGroup/details/sections/ApplicationManagerSection.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+
+import { CollapsibleSection } from '@spinnaker/core';
+import { ICloudFoundryServerGroupDetailsSectionProps } from './ICloudFoundryServerGroupDetailsSectionProps';
+
+export class ApplicationManagerSection extends React.Component<ICloudFoundryServerGroupDetailsSectionProps> {
+  constructor(props: ICloudFoundryServerGroupDetailsSectionProps) {
+    super(props);
+  }
+
+  public render(): JSX.Element {
+    const { serverGroup } = this.props;
+    return (
+      <>
+        {serverGroup.appsManagerUri && (
+          <CollapsibleSection heading="Apps Manager" defaultExpanded={true}>
+            <div>
+              <a href={serverGroup.appsManagerUri} target="_blank">
+                Apps Manager
+              </a>
+            </div>
+          </CollapsibleSection>
+        )}
+      </>
+    );
+  }
+}
