+++ app/scripts/modules/cloudfoundry/src/serverGroup/details/sections/ICloudFoundryServerGroupDetailsSectionProps.ts
@@ -1,0 +1,7 @@
+import { IServerGroupDetailsSectionProps } from '@spinnaker/core';
+
+import { ICloudFoundryServerGroup } from 'cloudfoundry/domain';
+
+export interface ICloudFoundryServerGroupDetailsSectionProps extends IServerGroupDetailsSectionProps {
+  serverGroup: ICloudFoundryServerGroup;
+}
