+++ app/scripts/modules/cloudfoundry/src/serverGroup/details/sections/PackageSection.tsx
@@ -1,0 +1,28 @@
+import * as React from 'react';
+
+import { CollapsibleSection } from '@spinnaker/core';
+
+import { ICloudFoundryServerGroupDetailsSectionProps } from './ICloudFoundryServerGroupDetailsSectionProps';
+
+export class PackageSection extends React.Component<ICloudFoundryServerGroupDetailsSectionProps> {
+  constructor(props: ICloudFoundryServerGroupDetailsSectionProps) {
+    super(props);
+  }
+
+  public render(): JSX.Element {
+    const { serverGroup } = this.props;
+    return (
+      <>
+        {serverGroup.droplet &&
+          serverGroup.droplet.sourcePackage && (
+            <CollapsibleSection heading="Package" defaultExpanded={true}>
+              <dl className="dl-horizontal dl-flex">
+                <dt>Checksum</dt>
+                <dd>{serverGroup.droplet.sourcePackage.checksum}</dd>
+              </dl>
+            </CollapsibleSection>
+          )}
+      </>
+    );
+  }
+}
