+++ app/scripts/modules/cloudfoundry/src/serverGroup/details/sections/index.ts
@@ -1,1 +1,8 @@
-export * from './cloudFoundryInfoDetailsSection';
+export * from './ServerGroupInformationSection';
+export * from './ApplicationManagerSection';
+export * from './MetricsSection';
+export * from './ServerGroupSizingSection';
+export * from './HealthCheckSection';
+export * from './PackageSection';
+export * from './BoundServicesSection';
+export * from './EvironmentVariablesSection';
